+++ sneak/config.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 
 frame_len = .1
 
@@ -19,7 +20,7 @@ food_values = {
 game_sizes = {
     's': (25, 20),
     'm': (50, 40),
-    'l': (80, 40),
+    'l': (80, 50),
 }
 
 initial_size = 4