+++ sneak/theme.py
@@ -1,35 +1,32 @@
+# -*- coding: utf-8 -*-
 
 import curses
-import stage
 
-colors_map = {}
-theme = None
-default_color = None
+from stage import stage
 
 
-def init():
-    global theme, colors_map, default_color
+class Theme:
 
-    theme = stage.chosen_theme
-    colors_map = get_colors_map()
-    default_color = theme['colors']['default']
+    colors_map = {}
+    theme = None
 
+    def __init__(self):
+        self.theme = stage.chosen_theme
+        self.colors_map = self.get_colors_map()
 
-def get_color(key):
-    return curses.color_pair(colors_map.get(key, 0))
+    def get_color(self, key):
+        return curses.color_pair(self.colors_map.get(key, 0))
 
+    def get_tile(self, key):
+        return self.theme['tiles'].get(key, ' ')
 
-def get_tile(key):
-    return theme['tiles'].get(key, ' ')
+    def get_colors_map(self):
+        out = {}
+        i = 1
 
+        for col in self.theme['colors'].iteritems():
+            curses.init_pair(i, col[1][0], col[1][1])
+            out[col[0]] = i
+            i += 1
 
-def get_colors_map():
-    out = {}
-
-    i = 1
-    for col in theme['colors'].iteritems():
-        curses.init_pair(i, col[1][0], col[1][1])
-        out[col[0]] = i
-        i += 1
-
-    return out
+        return out